refactor: move add command to add/
