+++ src/add/add-display.ts
@@ -1,9 +1,9 @@
 import pc from 'picocolors';
-import { formatList } from './lib/utils.ts';
-import { agents, getUniversalAgents, isUniversalAgent } from './lib/agents/index.ts';
-import type { AgentType } from './lib/types.ts';
-import type { InstallMode } from './lib/install/index.ts';
-import type { AuditResponse, PartnerAudit } from './lib/telemetry.ts';
+import { formatList } from '../lib/utils.ts';
+import { agents, getUniversalAgents, isUniversalAgent } from '../lib/agents/index.ts';
+import type { AgentType } from '../lib/types.ts';
+import type { InstallMode } from '../lib/install/index.ts';
+import type { AuditResponse, PartnerAudit } from '../lib/telemetry.ts';
 
 // ─── Security Advisory ───
