+++ src/add/add-prompt.test.ts
@@ -1,16 +1,16 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { promptForAgents } from './add.js';
-import * as skillLock from './lib/lock/skill-lock.js';
-import * as searchMultiselectModule from './lib/prompts/search-multiselect.js';
+import * as skillLock from '../lib/lock/skill-lock.js';
+import * as searchMultiselectModule from '../lib/prompts/search-multiselect.js';
 
 // Mock dependencies
-vi.mock('./lib/lock/skill-lock.js');
-vi.mock('./lib/prompts/search-multiselect.js');
-vi.mock('./lib/telemetry.js', () => ({
+vi.mock('../lib/lock/skill-lock.js');
+vi.mock('../lib/prompts/search-multiselect.js');
+vi.mock('../lib/telemetry.js', () => ({
   setVersion: vi.fn(),
   track: vi.fn(),
 }));
-vi.mock('../package.json', () => ({
+vi.mock('../../package.json', () => ({
   default: { version: '1.0.0' },
 }));
