+++ src/add/index.ts
@@ -1,0 +1,3 @@
+export { runAdd, initTelemetry } from './add.ts';
+export { parseAddOptions, type AddOptions } from './add-options.ts';
+export { promptForAgents } from './add-agents.ts';
